requirements.txt

paddlepaddle
paddleocr
transformers
torch
pandas
Pillow
